move base-classes/funcs from _browser.py back to _figure.py to facilitate review
