+++ mne/viz/raw.py
@@ -203,7 +203,8 @@ def plot_raw(raw, events=None, duration=10.0, start=0.0, n_channels=20,
     ``True``. This flag can be toggled by pressing 'd'.
     """
     from ..io.base import BaseRaw
-    from ._browser import _get_browser, get_browser_backend
+    from ._figure import get_browser_backend
+    from ._figure import _get_browser
 
     info = raw.info.copy()
     sfreq = info['sfreq']
@@ -439,7 +440,7 @@ def plot_raw_psd(raw, fmin=0, fmax=np.inf, tmin=None, tmax=None, proj=False,
     fig : instance of Figure
         Figure with frequency spectra of the data channels.
     """
-    from ._figure import _psd_figure
+    from ._mpl_figure import _psd_figure
     # handle FFT
     if n_fft is None:
         if tmax is None or not np.isfinite(tmax):
